Small fix to relocation table.
Currently, there seems to be a data type mismatch between the outputs of some codelets and the inputs of others. For example, the output of a SIMD operation might be 32-bit but then its loaded from DRAM into systolic at 8-bit. Need to investigate how to make this more clear.